refactor(common/device): move listApps entrypoint to listAppsUsecase
