+++ apps/cli/src/commands/appUninstallAll.ts
@@ -4,7 +4,10 @@ import { mergeMap, filter, map } from "rxjs/operators";
 import { withDevice } from "@ledgerhq/live-common/hw/deviceAccess";
 import getDeviceInfo from "@ledgerhq/live-common/hw/getDeviceInfo";
 import { reducer, runAll } from "@ledgerhq/live-common/apps/index";
-import { listApps, execWithTransport } from "@ledgerhq/live-common/apps/hw";
+import {
+  listAppsUseCase,
+  execWithTransport,
+} from "@ledgerhq/live-common/device/use-cases/listAppsUseCase";
 import { command as uninstallAllApps } from "@ledgerhq/live-common/hw/uninstallAllApps";
 import { deviceOpt } from "../scan";
 
@@ -25,7 +28,7 @@ export default {
           } else {
             return from(getDeviceInfo(t)).pipe(
               mergeMap(deviceInfo =>
-                listApps(t, deviceInfo).pipe(
+                listAppsUseCase(t, deviceInfo).pipe(
                   filter(e => e.type === "result"),
                   map((e: any) =>
                     reducer(e.result, {
