+++ apps/cli/src/commands/appsInstallAll.ts
@@ -4,7 +4,10 @@ import { mergeMap, filter, map } from "rxjs/operators";
 import { withDevice } from "@ledgerhq/live-common/hw/deviceAccess";
 import getDeviceInfo from "@ledgerhq/live-common/hw/getDeviceInfo";
 import { initState, reducer, runAll } from "@ledgerhq/live-common/apps/index";
-import { listApps, execWithTransport } from "@ledgerhq/live-common/apps/hw";
+import {
+  listAppsUseCase,
+  execWithTransport,
+} from "@ledgerhq/live-common/device/use-cases/listAppsUseCase";
 import { deviceOpt } from "../scan";
 export default {
   description: "test script to install and uninstall all apps",
@@ -18,7 +21,7 @@ export default {
       const exec = execWithTransport(t);
       return from(getDeviceInfo(t)).pipe(
         mergeMap(deviceInfo =>
-          listApps(t, deviceInfo).pipe(
+          listAppsUseCase(t, deviceInfo).pipe(
             filter(e => e.type === "result"),
             map((e: any) =>
               e.result.appsListNames.reduce(
