+++ apps/ledger-live-desktop/src/renderer/Default.tsx
@@ -40,7 +40,7 @@ import VaultSignerBanner from "~/renderer/components/VaultSignerBanner";
 import { hasCompletedOnboardingSelector } from "~/renderer/reducers/settings";
 import { updateIdentify } from "./analytics/segment";
 import { useFeature, FeatureToggle } from "@ledgerhq/live-common/featureFlags/index";
-import { enableListAppsV2 } from "@ledgerhq/live-common/apps/hw";
+import { enableListAppsV2 } from "@ledgerhq/live-common/device/use-cases/listAppsUseCase";
 import {
   useFetchCurrencyAll,
   useFetchCurrencyFrom,
