--- libs/ledger-live-common/src/device/use-cases/listAppsV2UseCase.ts
@@ -1,12 +1,0 @@
-import { Observable } from "rxjs";
-import Transport from "@ledgerhq/hw-transport";
-import { DeviceInfo } from "@ledgerhq/types-live";
-import { listAppsV2 } from "../../apps/listApps/v2";
-import { ListAppsEvent } from "../../apps";
-
-export function listAppsV2UseCase(
-  transport: Transport,
-  deviceInfo: DeviceInfo,
-): Observable<ListAppsEvent> {
-  return listAppsV2({transport, deviceInfo, deviceProxyModel: getEnv("DEVICE_PROXY_MODEL") as DeviceModelId});
-}
